Move IndexManager code into FieldIndexManagementApiImpl
